refactor(accounting): update import path for accountOptions
